requirements.txt
# 기본 라이브러리 (버전은 필요에 따라 조정)
streamlit>=1.30.0
pandas>=1.5.0
pytz>=2023.3
openpyxl>=3.0.10
reportlab>=4.0.0

# Google Drive 연동에 필요한 라이브러리
google-api-python-client>=2.80.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# PDF를 이미지로 변환하기 위한 라이브러리
pdf2image>=1.16.0
Pillow>=9.0.0

# MMS 게이트웨이 연동 시 HTTP 요청을 위한 라이브러리 (mms_utils.py 예시에서 사용)
requests>=2.25.0
